+++ src/main/java/blackjack/domain/player/Player.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
+import blackjack.domain.GameOutcome;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
 import java.util.List;
