+++ src/main/java/blackjack/domain/player/Players.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 import blackjack.domain.card.Card;
 import blackjack.dto.PlayerInfo;
